# api/openapi.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: |
    Lean Go backend for the core fitness tracking loop: users, exercise library,
    workouts, workout exercises, workout sets, meals, and weight entries.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/meals/recent:
    get:
      tags: [Meals]
      summary: Get recently logged meals
      operationId: getRecentMeals
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
          description: Number of days back to look.
      responses:
        "200":
          description: List of recent meals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMeals"
        "401":
          $ref: "#/components/responses/Error"
  /v1/meals/{id}/clone:
    post:
      tags: [Meals]
      summary: Clone an existing meal
      operationId: cloneMeal
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Meal cloned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/foods/recent:
    get:
      tags: [Foods]
      summary: Get recently used foods
      operationId: getRecentFoods
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of recently used foods.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFoods"
        "401":
          $ref: "#/components/responses/Error"
...
  /v1/foods/{id}/favorite:
    post:
      tags: [Foods]
      summary: Favorite a food
      operationId: favoriteFood
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Food favorited.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteFood"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Foods]
      summary: Unfavorite a food
      operationId: unfavoriteFood
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "204":
          description: Food unfavorited.
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/users/{user_id}/favorites:
    get:
      tags: [Foods]
      summary: Get user's favorite foods
      operationId: getUserFavorites
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: List of favorite foods.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFavoriteFoods"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /v1/users/{user_id}/nutrition-targets:
    get:
      tags: [Summaries]
      summary: Get user nutrition targets
      operationId: getUserNutritionTargets
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Nutrition targets.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    UserID:
      in: path
      name: user_id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [title, status, error]
      properties:
        type:
          type: string
          description: Problem type URI. Validation errors use a dedicated type; generic errors use `about:blank`.
          example: about:blank
        title:
          type: string
          description: Short, human-readable summary of the error category.
          example: Bad Request
        status:
          type: integer
          description: HTTP status code for the error response.
          example: 400
        detail:
          type: string
          description: Human-readable explanation of the specific error instance.
          example: invalid email or password
        error:
          type: string
          description: Backward-compatible alias of `detail`.
          example: invalid email or password
        errors:
          type: object
          description: Field-specific validation messages when the request fails validation.
          additionalProperties:
            type: string
          example:
            email: email is required
            password: password must be at least 8 characters
...
    FavoriteFood:
      type: object
      required: [user_id, food_id]
      properties:
        user_id:
          type: string
          format: uuid
        food_id:
          type: string
          format: uuid
        food:
          $ref: "#/components/schemas/Food"
        created_at:
          type: string
          format: date-time
    Food:
      type: object
      required: [id, name, serving_unit, calories]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        brand:
          type: string
        category:
          type: string
        source:
          type: string
        serving_size:
          type: number
        serving_unit:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbohydrates:
          type: number
        fat:
          type: number
        fiber:
          type: number
        sugar:
          type: number
        sodium:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Meal:
      type: object
      required:
        - id
        - user_id
        - meal_type
        - date
        - notes
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        meal_type:
          type: string
        date:
          type: string
          format: date-time
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaginationMetadata:
      type: object
      required: [page, limit, total_count, total_pages, has_next]
      properties:
        page:
          type: integer
          description: Current page number.
        limit:
          type: integer
          description: Maximum number of records requested per page.
        total_count:
          type: integer
          description: Total number of records matching the query.
        total_pages:
          type: integer
          description: Total number of pages available for the current limit.
        has_next:
          type: boolean
          description: Whether another page of results exists after the current page.
...
    PaginatedMeals:
      type: object
      required: [data, metadata]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
    PaginatedFoods:
      type: object
      required: [data, metadata]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Food"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
    PaginatedFavoriteFoods:
      type: object
      required: [data, metadata]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteFood"
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT